I remember why vault didn't use base now.
We need to give the binary CAP_IPC_LOCK=+ep to run it as a non-root
user, so we need (a) libcap installed, and (b) to start in the
final stage as root so we can invoke setcap.

This is mostly a big copypaste of the base image with some tweaks,
might be able to dedupe it nicely later...

Also fix the address env var, which was just nonsense. Go me.